requirements.txt
streamlit>=1.28.0
pandas>=2.0.0
gspread>=5.12.0
folium>=0.15.0
streamlit-folium>=0.19.0
plotly>=5.18.0
google-auth>=2.23.0
requests>=2.31.0